+++ test/server/conf.js
@@ -1,0 +1,3 @@
+module.exports = {
+    testPort: "9876"
+};