+++ test/unit/pack.test.js
@@ -1,22 +1,25 @@
 "use strict";
 
+// Set test environment
+process.env.NODE_ENV = "test";
+
 const chai = require("chai");
 const expect = chai.expect;
 const fs = require("fs");
 const path = require("path");
 
-const pack = require("../lib/pack");
-const share = require("./data/share");
+const pack = require("../../lib/pack");
+const share = require("../data/share");
 
-describe("pack", () => {
-    let testFolder = path.join(__dirname, "data", "pack"),
-        srcFolder = path.join(__dirname, "data", "bowerPkgExample");
+describe("pack", function() {
+    const testFolder = path.join(__dirname, "..", "data", "pack");
+    const srcFolder = path.join(__dirname, "..", "data", "bowerPkgExample");
 
-    before((done) => {
+    before(function(done) {
         share.createTestFolder(testFolder, done);
     });
 
-    it("first creation", (done) => {
+    it("first creation", function(done) {
         pack(srcFolder, testFolder, false, (err, data) => {
             try {
                 expect(err).to.be.null;
@@ -32,7 +35,7 @@ describe("pack", () => {
         });
     });
 
-    it("overwrite", (done) => {
+    it("overwrite", function(done) {
         pack(srcFolder, testFolder, true, (err, data) => {
             try {
                 expect(err).to.be.null;
@@ -48,7 +51,7 @@ describe("pack", () => {
         });
     });
 
-    after((done) => {
+    after(function(done) {
         share.deleteTestFolder(testFolder, done);
     });
 });